requirements.txt
flask==3.0.0
flask-cors==4.0.0
librosa==0.10.1
openai-whisper==20231117
moviepy==1.0.3
ffmpeg-python==0.2.0
numpy==1.24.3
scipy==1.11.3
transformers==4.36.0